requirements.txt
# NEXUS - Maternal-Neonatal Care Platform
# Python Dependencies for MedGemma Impact Challenge

# Core ML/AI
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
safetensors>=0.4.0
sentencepiece>=0.1.99

# HAI-DEF Models
google-cloud-aiplatform>=1.38.0

# Model Optimization (Edge Deployment)
onnx>=1.15.0
onnxruntime>=1.16.0
optimum>=1.15.0

# Audio Processing (HeAR / Cry Analysis)
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Image Processing (MedSigLIP / Anemia & Jaundice)
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
einops>=0.7.0

# Demo Applications
streamlit>=1.28.0
gradio>=4.0.0
plotly>=5.18.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
rich>=13.0.0

# Dataset Download
kaggle>=1.5.0

# Jupyter & Visualization
jupyter>=1.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
